requirements.txt
crewai
crewai-tools
langchain
langchain-openai
python-dotenv
pypdf
fastapi
openai